[ IMPORTANT ][ Shadows ][ Dist-to-occluder-search ][ Fix ][ Optimization ][ Refactor ]
Simplified dist-to-occluder-search as it no more has to support such wild variety of values in the texture (because of splitting the shadow processing into 3 layers).

Improved dist-to-occluder-search to give more smooth results, without jagged edges etc. Added radial weighting in dist-to-occluder-search shader - to improve blending of nearby regions (plus early rejecting samples without sampling the texture).

Added sampling of different input mipmap levels during dist-to-occluder-search for hard/medium/soft shadows. It's better than skipping pixels, as it ensures smooth output image rather than undersampling artefacts and patterns.

Added a setting to control final-dist-to-occluder texture diemnsions for hard/medium/soft shadows - important as it directly influences the number of executed shader threads (caching hides the benefits though).

Fixed a bug in mipmap-generation-with-sample-rejection shader - it was outputting negative values because of missing max(). This was causing jagged edges in dist-to-occluder mipmaps.

Fixed another bug in generate-mipmaps-with-sample-rejection shader - it was outputting zero if all samples were rejected instead of returning the original value - that caused background switch from white to black and increasing darkening of edges in dist-to-occluder mipmaps.